Выполнено задание #18 Больше модулей
